all specifics are now handled by fa18Handler
